refactor(adapter): put a cleaner way for inheritance
