+++ src/adapters/LdsTestWireAdapter.js
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 
-import { TestWireAdapterTemplate } from "./TestWireAdapter";
+import { generateAdapter } from "./adapter";
 
 function buildErrorObject({ body, status, statusText }) {
     if (status && (status < 400 || status > 599)) {
@@ -29,33 +29,28 @@ function buildErrorObject({ body, status, statusText }) {
     };
 }
 
-class LdsTestWireAdapterTemplate extends TestWireAdapterTemplate {
-    static emit(value, filterFn) {
-        super.emit({ data: value, error: undefined }, filterFn);
-    }
-
-    static emitError(errorOptions, filterFn) {
-        const err = buildErrorObject(errorOptions || {});
-
-        super.emit({ data: undefined, error: err }, filterFn);
-    }
-
-    static error(body, status, statusText) {
-        const err = buildErrorObject({ body, status, statusText });
-
-        super.emit({ data: undefined, error: err });
-    }
-
-    constructor(dataCallback) {
-        super(dataCallback);
-
-        this.emit({ data: undefined, error: undefined })
-    }
-}
-
 export function buildLdsTestWireAdapter() {
-    return class LdsTestWireAdapter extends LdsTestWireAdapterTemplate {
-        static _lastConfig = null;
-        static _wireInstances = new Set();
-    }
+    return class LdsTestWireAdapter extends generateAdapter() {
+        static emit(value, filterFn) {
+            super.emit({ data: value, error: undefined }, filterFn);
+        }
+
+        static emitError(errorOptions, filterFn) {
+            const err = buildErrorObject(errorOptions || {});
+
+            super.emit({ data: undefined, error: err }, filterFn);
+        }
+
+        static error(body, status, statusText) {
+            const err = buildErrorObject({ body, status, statusText });
+
+            super.emit({ data: undefined, error: err });
+        }
+
+        constructor(dataCallback) {
+            super(dataCallback);
+
+            this.emit({ data: undefined, error: undefined })
+        }
+    };
 }
